requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
drf_spectacular==0.27.2
mysqlclient==2.2.4
PyJWT==2.9.0
pytz==2024.1
